# CUDA: MMQ support for iq4_nl, iq4_xs