… ggml/src/ggml-cuda/template-instances/generate_cu_files.py
… ggml/src/ggml-cuda/template-instances/mmq-instance-iq4_nl.cu
… ggml/src/ggml-cuda/template-instances/mmq-instance-iq4_xs.cu