This PR adds MMQ support for iq4_nl and iq4_xs. The data is loaded, converted to 8 bit, and written to shared memory. Because this is the same strategy as for q5_0 the same code can be re-used except for the part that loads the data. 

The other iq data types have issues with shared memory limits with the current MMQ code; it will need a refactor that allows setting the tile size in k direction as a configurable parameter. Presumably due to Ampere/Ada Lovelace consumer cards having 50% more shared memory than Turing this will also mean that for optimal performance there would then need to be different template instances for Turing and Ampere+. For A100s/H100s which have even more shared memory you would in principle also need different configurations but I am not interested in working on that hardware since I will not be able to afford it anyways.

Since I am already working on the MMQ code I replaced instances of `get_int_from_int8` with the refactored and implified variants that accept void pointers.

<details>
<summary>Performance</summary>

| Model                      | GPU      |   Microbatch size | Test   |   t/s master |   t/s cuda-iq-mmq-2 |   Speedup |
|:---------------------------|:---------|------------------:|:-------|-------------:|--------------------:|----------:|
| llama 8B IQ4_NL - 4.5 bpw  | RX 6800  |                16 | pp2048 |        67.10 |              236.65 |      3.53 |
| llama 8B IQ4_NL - 4.5 bpw  | RX 6800  |                32 | pp2048 |       120.16 |              328.50 |      2.73 |
| llama 8B IQ4_NL - 4.5 bpw  | RX 6800  |                63 | pp2048 |       183.65 |              396.08 |      2.16 |
| llama 8B IQ4_NL - 4.5 bpw  | RX 6800  |               128 | pp2048 |       337.26 |              509.96 |      1.51 |
| llama 8B IQ4_NL - 4.5 bpw  | RX 6800  |               256 | pp2048 |       519.73 |              606.06 |      1.17 |
| llama 8B IQ4_NL - 4.5 bpw  | RX 6800  |               512 | pp2048 |       577.88 |              611.98 |      1.06 |
| llama 8B IQ4_NL - 4.5 bpw  | RX 6800  |              1024 | pp2048 |       594.07 |              687.13 |      1.16 |
| llama 8B IQ4_NL - 4.5 bpw  | RX 6800  |              2048 | pp2048 |       581.07 |              625.90 |      1.08 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 3090 |                16 | pp2048 |       342.89 |             1046.94 |      3.05 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 3090 |                32 | pp2048 |       646.45 |             1713.55 |      2.65 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 3090 |                63 | pp2048 |      1211.66 |             2552.69 |      2.11 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 3090 |               128 | pp2048 |      2138.26 |             3225.39 |      1.51 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 3090 |               256 | pp2048 |      3149.48 |             3621.78 |      1.15 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 3090 |               512 | pp2048 |      3718.51 |             3762.09 |      1.01 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 3090 |              1024 | pp2048 |      4357.64 |             3736.55 |      0.86 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 3090 |              2048 | pp2048 |      4376.58 |             3621.46 |      0.83 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 4090 |                16 | pp2048 |       491.99 |             1950.90 |      3.97 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 4090 |                32 | pp2048 |       970.65 |             3398.93 |      3.50 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 4090 |                63 | pp2048 |      1754.37 |             5423.63 |      3.09 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 4090 |               128 | pp2048 |      3432.75 |             7492.11 |      2.18 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 4090 |               256 | pp2048 |      5706.46 |             9627.65 |      1.69 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 4090 |               512 | pp2048 |      7641.30 |            10338.84 |      1.35 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 4090 |              1024 | pp2048 |      8857.21 |            10008.78 |      1.13 |
| llama 8B IQ4_NL - 4.5 bpw  | RTX 4090 |              2048 | pp2048 |      8729.23 |             9000.75 |      1.03 |
| llama 8B IQ4_NL - 4.5 bpw  | P40      |                16 | pp2048 |        48.57 |              249.63 |      5.14 |
| llama 8B IQ4_NL - 4.5 bpw  | P40      |                32 | pp2048 |        95.67 |              418.39 |      4.37 |
| llama 8B IQ4_NL - 4.5 bpw  | P40      |                63 | pp2048 |       167.29 |              557.72 |      3.33 |
| llama 8B IQ4_NL - 4.5 bpw  | P40      |               128 | pp2048 |       212.61 |              681.91 |      3.21 |
| llama 8B IQ4_NL - 4.5 bpw  | P40      |               256 | pp2048 |       324.19 |              770.01 |      2.38 |
| llama 8B IQ4_NL - 4.5 bpw  | P40      |               512 | pp2048 |       446.36 |              807.85 |      1.81 |
| llama 8B IQ4_NL - 4.5 bpw  | P40      |              1024 | pp2048 |       511.27 |              797.13 |      1.56 |
| llama 8B IQ4_NL - 4.5 bpw  | P40      |              2048 | pp2048 |       523.17 |              765.75 |      1.46 |
| llama 8B IQ4_XS - 4.25 bpw | RX 6800  |                16 | pp2048 |        67.23 |              236.07 |      3.51 |
| llama 8B IQ4_XS - 4.25 bpw | RX 6800  |                32 | pp2048 |       120.34 |              331.87 |      2.76 |
| llama 8B IQ4_XS - 4.25 bpw | RX 6800  |                63 | pp2048 |       183.93 |              397.62 |      2.16 |
| llama 8B IQ4_XS - 4.25 bpw | RX 6800  |               128 | pp2048 |       338.18 |              512.38 |      1.52 |
| llama 8B IQ4_XS - 4.25 bpw | RX 6800  |               256 | pp2048 |       520.16 |              610.81 |      1.17 |
| llama 8B IQ4_XS - 4.25 bpw | RX 6800  |               512 | pp2048 |       578.30 |              615.09 |      1.06 |
| llama 8B IQ4_XS - 4.25 bpw | RX 6800  |              1024 | pp2048 |       596.10 |              692.03 |      1.16 |
| llama 8B IQ4_XS - 4.25 bpw | RX 6800  |              2048 | pp2048 |       580.27 |              630.13 |      1.09 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 3090 |                16 | pp2048 |       345.80 |             1051.77 |      3.04 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 3090 |                32 | pp2048 |       654.85 |             1701.79 |      2.60 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 3090 |                63 | pp2048 |      1227.19 |             2525.81 |      2.06 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 3090 |               128 | pp2048 |      2153.51 |             3166.41 |      1.47 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 3090 |               256 | pp2048 |      3150.56 |             3497.60 |      1.11 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 3090 |               512 | pp2048 |      3714.67 |             3629.68 |      0.98 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 3090 |              1024 | pp2048 |      4279.30 |             3654.98 |      0.85 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 3090 |              2048 | pp2048 |      4264.30 |             3567.27 |      0.84 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 4090 |                16 | pp2048 |       493.56 |             1970.30 |      3.99 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 4090 |                32 | pp2048 |       976.59 |             3417.54 |      3.50 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 4090 |                63 | pp2048 |      1766.28 |             5407.20 |      3.06 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 4090 |               128 | pp2048 |      3471.43 |             7390.92 |      2.13 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 4090 |               256 | pp2048 |      5724.22 |             9443.06 |      1.65 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 4090 |               512 | pp2048 |      7643.08 |            10119.12 |      1.32 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 4090 |              1024 | pp2048 |      8883.77 |             9814.66 |      1.10 |
| llama 8B IQ4_XS - 4.25 bpw | RTX 4090 |              2048 | pp2048 |      8770.67 |             8813.15 |      1.00 |
| llama 8B IQ4_XS - 4.25 bpw | P40      |                16 | pp2048 |        48.33 |              258.02 |      5.34 |
| llama 8B IQ4_XS - 4.25 bpw | P40      |                32 | pp2048 |        95.25 |              418.44 |      4.39 |
| llama 8B IQ4_XS - 4.25 bpw | P40      |                63 | pp2048 |       166.63 |              567.39 |      3.41 |
| llama 8B IQ4_XS - 4.25 bpw | P40      |               128 | pp2048 |       212.48 |              687.74 |      3.24 |
| llama 8B IQ4_XS - 4.25 bpw | P40      |               256 | pp2048 |       324.41 |              772.64 |      2.38 |
| llama 8B IQ4_XS - 4.25 bpw | P40      |               512 | pp2048 |       451.42 |              812.53 |      1.80 |
| llama 8B IQ4_XS - 4.25 bpw | P40      |              1024 | pp2048 |       512.68 |              802.59 |      1.57 |
| llama 8B IQ4_XS - 4.25 bpw | P40      |              2048 | pp2048 |       524.43 |              771.65 |      1.47 |

</details>