@ExtReMLapin 2024-07-03
Funny to see that on a rtx 4090, a higher microbatch size doesn't mean higher speed