@JohannesGaessler 2024-07-03
This has nothing to do with MMQ though; the MMQ runtime still goes down by ~3% if you increase the batch size from 512 to 2048. The problem is instead inefficient masking in the FlashAttention kernel where larger batch sizes lead to iteration over more values that are masked out anyways. cuBLAS has the same problem but in return suffers less from dequantization overhead at large batch sizes.